Move confirmation action to the confirmation controller
