+++ spec/controllers/user_controller_spec.rb
@@ -31,234 +31,6 @@ describe UserController do
 
   end
 
-  describe 'GET confirm' do
-
-    context 'if the post redirect cannot be found' do
-
-      it 'renders bad_token' do
-        get :confirm, :email_token => ''
-        expect(response).to render_template(:bad_token)
-      end
-
-    end
-
-    context 'the post redirect circumstance is login_as' do
-
-      before :each do
-        @user = FactoryGirl.create(:user, :email_confirmed => false)
-        @post_redirect =
-          PostRedirect.
-            create(:uri => '/', :user => @user, :circumstance => 'login_as')
-
-        get :confirm, :email_token => @post_redirect.email_token
-      end
-
-      it 'confirms the post redirect user' do
-        expect(@user.reload.email_confirmed).to eq(true)
-      end
-
-      it 'logs in as the post redirect user' do
-        expect(session[:user_id]).to eq(@user.id)
-      end
-
-      it 'sets the user_circumstance to login_as' do
-        expect(session[:user_circumstance]).to eq('login_as')
-      end
-
-      it 'redirects to the post redirect uri' do
-        expect(response).to redirect_to('/?post_redirect=1')
-      end
-
-    end
-
-    context 'the post redirect circumstance is change_password' do
-
-      before :each do
-        @user = FactoryGirl.create(:user)
-        @post_redirect =
-          PostRedirect.create(:uri => edit_password_change_path,
-                              :user => @user,
-                              :circumstance => 'change_password')
-
-        get :confirm, :email_token => @post_redirect.email_token
-      end
-
-      it 'sets the change_password_post_redirect_id session key' do
-         expect(session[:change_password_post_redirect_id]).
-           to eq(@post_redirect.id)
-      end
-
-      it 'does not log the user in' do
-        expect(session[:user_id]).to eq(nil)
-      end
-
-      it 'logs out a user who does not own the post redirect' do
-        logged_in_user = FactoryGirl.create(:user)
-        @user = FactoryGirl.create(:user, :email_confirmed => false)
-        @post_redirect =
-          PostRedirect.create(:uri => edit_password_change_path,
-                              :user => @user,
-                              :circumstance => 'change_password')
-
-        session[:user_id] = logged_in_user.id
-        get :confirm, :email_token => @post_redirect.email_token
-
-        expect(session[:user_id]).to be_nil
-      end
-
-      it 'does not log out a user if they own the post redirect' do
-        @user = FactoryGirl.create(:user, :email_confirmed => false)
-        @post_redirect =
-          PostRedirect.create(:uri => edit_password_change_path,
-                              :user => @user,
-                              :circumstance => 'change_password')
-
-        session[:user_id] = @user.id
-        get :confirm, :email_token => @post_redirect.email_token
-
-        expect(session[:user_id]).to eq(@user.id)
-        expect(assigns[:user]).to eq(@user)
-      end
-
-      it 'does not confirm an unconfirmed user' do
-        @user = FactoryGirl.create(:user, :email_confirmed => false)
-        @post_redirect =
-          PostRedirect.create(:uri => edit_password_change_path,
-                              :user => @user,
-                              :circumstance => 'change_password')
-
-        get :confirm, :email_token => @post_redirect.email_token
-
-        expect(@user.reload.email_confirmed).to eq(false)
-      end
-
-      it 'sets the user_circumstance to change_password' do
-        expect(session[:user_circumstance]).to eq('change_password')
-      end
-
-      it 'redirects to the post redirect uri' do
-        expect(response).
-          to redirect_to('/profile/change_password?post_redirect=1')
-      end
-
-    end
-
-    context 'if the currently logged in user is an admin' do
-
-      before :each do
-        @admin = FactoryGirl.create(:admin_user)
-        @user = FactoryGirl.create(:user, :email_confirmed => false)
-        @post_redirect = PostRedirect.create(:uri => '/', :user => @user)
-
-        session[:user_id] = @admin.id
-        get :confirm, :email_token => @post_redirect.email_token
-      end
-
-      it 'does not confirm the post redirect user' do
-        expect(@user.reload.email_confirmed).to eq(false)
-      end
-
-      it 'stays logged in as the admin user' do
-        expect(session[:user_id]).to eq(@admin.id)
-      end
-
-      it 'sets the user_circumstance to normal' do
-        expect(session[:user_circumstance]).to eq('normal')
-      end
-
-      it 'redirects to the post redirect uri' do
-        expect(response).to redirect_to('/?post_redirect=1')
-      end
-
-    end
-
-    context 'if the currently logged in user is not an admin and owns the post redirect' do
-
-      before :each do
-        @user = FactoryGirl.create(:user, :email_confirmed => false)
-        @post_redirect = PostRedirect.create(:uri => '/', :user => @user)
-
-        session[:user_id] = @user.id
-        get :confirm, :email_token => @post_redirect.email_token
-      end
-
-      it 'confirms the post redirect user' do
-        expect(@user.reload.email_confirmed).to eq(true)
-      end
-
-      it 'stays logged in as the user' do
-        expect(session[:user_id]).to eq(@user.id)
-      end
-
-      it 'sets the user_circumstance to normal' do
-        expect(session[:user_circumstance]).to eq('normal')
-      end
-
-      it 'redirects to the post redirect uri' do
-        expect(response).to redirect_to('/?post_redirect=1')
-      end
-
-    end
-
-    context 'if the currently logged in user is not an admin and does not own the post redirect' do
-
-      before :each do
-        @current_user = FactoryGirl.create(:user)
-        @user = FactoryGirl.create(:user, :email_confirmed => false)
-        @post_redirect = PostRedirect.create(:uri => '/', :user => @user)
-
-        session[:user_id] = @current_user.id
-        get :confirm, :email_token => @post_redirect.email_token
-      end
-
-      it 'confirms the post redirect user' do
-        expect(@user.reload.email_confirmed).to eq(true)
-      end
-
-      # FIXME: There's no reason this should be allowed
-      it 'gets logged in as the post redirect user' do
-        expect(session[:user_id]).to eq(@user.id)
-      end
-
-      it 'sets the user_circumstance to normal' do
-        expect(session[:user_circumstance]).to eq('normal')
-      end
-
-      it 'redirects to the post redirect uri' do
-        expect(response).to redirect_to('/?post_redirect=1')
-      end
-
-    end
-
-    context 'if there is no logged in user' do
-
-      before :each do
-        @user = FactoryGirl.create(:user, :email_confirmed => false)
-        @post_redirect = PostRedirect.create(:uri => '/', :user => @user)
-
-        get :confirm, :email_token => @post_redirect.email_token
-      end
-
-      it 'confirms the post redirect user' do
-        expect(@user.reload.email_confirmed).to eq(true)
-      end
-
-      it 'gets logged in as the post redirect user' do
-        expect(session[:user_id]).to eq(@user.id)
-      end
-
-      it 'sets the user_circumstance to normal' do
-        expect(session[:user_circumstance]).to eq('normal')
-      end
-
-      it 'redirects to the post redirect uri' do
-        expect(response).to redirect_to('/?post_redirect=1')
-      end
-
-    end
-
-  end
-
 end
 
 # TODO: Use route_for or params_from to check /c/ links better
@@ -823,60 +595,6 @@ describe UserController, "when changing email address" do
                              }
 
     expect(response).to render_template('signchangeemail_confirm')
-  end
-
-  it "should send confirmation email if you get all the details right" do
-    @user = users(:bob_smith_user)
-    session[:user_id] = @user.id
-
-    post :signchangeemail, { :signchangeemail => { :old_email => 'bob@localhost',
-                                                   :password => 'jonespassword', :new_email => 'newbob@localhost' },
-                             :submitted_signchangeemail_do => 1
-                             }
-
-    @user.reload
-    expect(@user.email).to eq('bob@localhost')
-    expect(@user.email_confirmed).to eq(true)
-
-    expect(response).to render_template('signchangeemail_confirm')
-
-    deliveries = ActionMailer::Base.deliveries
-    expect(deliveries.size).to  eq(1)
-    mail = deliveries[0]
-    expect(mail.body).to include("confirm that you want to change")
-    expect(mail.to).to eq([ 'newbob@localhost' ])
-
-    mail.body.to_s =~ /(http:\/\/.*(\/c\/(.*)))/
-    mail_url = $1
-    mail_path = $2
-    mail_token = $3
-
-    # Check confirmation URL works
-    session[:user_id] = nil
-    expect(session[:user_circumstance]).to eq(nil)
-    get :confirm, :email_token => mail_token
-    expect(session[:user_id]).to eq(users(:bob_smith_user).id)
-    expect(session[:user_circumstance]).to eq('change_email')
-    expect(response).to redirect_to(:controller => 'user', :action => 'signchangeemail', :post_redirect => 1)
-
-    # Would be nice to do a follow_redirect! here, but rspec-rails doesn't
-    # have one. Instead do an equivalent manually.
-    post_redirect = PostRedirect.find_by_email_token(mail_token)
-    expect(post_redirect.circumstance).to eq('change_email')
-    expect(post_redirect.user).to eq(users(:bob_smith_user))
-    expect(post_redirect.post_params).to eq({"submitted_signchangeemail_do"=>"1",
-                                         "action"=>"signchangeemail",
-                                         "signchangeemail"=>{
-                                           "old_email"=>"bob@localhost",
-                                         "new_email"=>"newbob@localhost"},
-                                         "controller"=>"user"})
-    post :signchangeemail, post_redirect.post_params
-
-    expect(response).to redirect_to(:controller => 'user', :action => 'show', :url_name => 'bob_smith')
-    expect(flash[:notice]).to match(/You have now changed your email address/)
-    @user.reload
-    expect(@user.email).to eq('newbob@localhost')
-    expect(@user.email_confirmed).to eq(true)
   end
 
   it "should send special 'already signed up' mail if you try to change your email to one already used" do
