+++ spec/controllers/users/sessions_controller_spec.rb
@@ -112,4 +112,5 @@ describe Users::SessionsController do
     end
 
   end
+
 end
