+++ spec/integration/alaveteli_dsl.rb
@@ -142,6 +142,14 @@ def last_holding_pen_mail
   InfoRequest.holding_pen_request.get_last_public_response.raw_email
 end
 
+def confirmation_url_from_email
+  deliveries = ActionMailer::Base.deliveries
+  expect(deliveries.size).to eq(1)
+  mail = deliveries.first
+  mail.body.to_s =~ /(http:\/\/.*\/c\/.*)/
+  $1
+end
+
 def cache_directories_exist?(request)
   cache_path = File.join(Rails.root, 'cache', 'views')
   paths = [File.join(cache_path, 'request', request.request_dirs)]
